requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langchain-community
pydantic
langchain-core 